Support more metacomments and introduce some validation effort
In case of errors, the composed toot is fed back into the editor,
including a message about what went wrong.
